+++ CHANGELOG.md
@@ -7,6 +7,10 @@
 * Transferring items via shift-clicking in the player's inventory during PvP is now disableable
 * If an inventory contains less stacks than have to be dropped, the game can optionally scan other inventories
 * Players can now optionally decide with `pvp spy [on|off]` whether they want to provide/receive proximity informations
+* Changed `pvplist` to show only players with PvP enabled by default
+* Added `pvplist all` which shows the whole list
+* Added `pvplist <maxEntryCount>` which shows the specified number of entries
+* The displayed player list in `pvplist` now shows the count of displayed entries
 
 ### LOTR compatibility:
 * Players now optionally drop their skulls when killed with a weapon with the headhunter modifier even if keepInventory is on